requirements.txt
python-Levenshtein==0.12.2
thefuzz==0.19.0

beautifulsoup4~=4.11.1
requests~=2.27.1
tqdm~=4.64.0

# {{< VulcanTear >}}
fire~=0.4.0
pytz~=2022.1
pyyaml~=6.0
gevent~=21.12.0
loguru~=0.6.0
aiohttp~=3.8.1
lxml~=4.8.0

# {{< TLS in TLS >}}
# urllib3==1.25.11

markupsafe==2.0.1
Flask
numpy
flask-bootstrap
pytz
